# recipedia-app/src/main/java/com/recipedia/dto/ReviewResponse.java
package com.recipedia.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class ReviewResponse {

    private Double rate;
    private String comment;
    private boolean ownFeedback;

}