# recipedia-ui/src/app/services/models/page-response-recipe-response.ts
/* tslint:disable */
/* eslint-disable */
import { RecipeResponse } from '../models/recipe-response';
export interface PageResponseRecipeResponse {
  content?: Array<RecipeResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}