# recipedia-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Recipedia OpenApi specification",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of Service",
    "contact": {
      "name": "rzn",
      "url": "https://rzndev.com",
      "email": "rosenberg2001fb@gmail.com"
    },
    "license": { "name": "License1", "url": "https://Random-license-url.com " },
    "version": "1.0"
  },
  "servers": [
    { "url": "http://localhost:8080/api/v1", "description": "Local ENV" },
    { "url": "https://rnzdev.com", "description": "Prod ENV" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/reviews": {
      "post": {
        "tags": ["Review"],
        "operationId": "saveReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReviewRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": { "schema": { "type": "integer", "format": "int64" } }
            }
          }
        }
      }
    },
    "/recipes": {
      "get": {
        "tags": ["Recipe"],
        "operationId": "findAllRecipes",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["KOREAN", "QUICK", "VEGAN", "BAKERY", "PASTA"]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseRecipeResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Recipe"],
        "operationId": "saveRecipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RecipeRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": { "schema": { "type": "integer", "format": "int64" } }
            }
          }
        }
      }
    },
    "/recipes/cover/{recipe-id}": {
      "post": {
        "tags": ["Recipe"],
        "operationId": "uploadRecipeStepImage",
        "parameters": [
          {
            "name": "recipe-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "stepNumber",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": ["file"],
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegistrationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/auth/authenticate": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthenticationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/recipe/{recipe-id}": {
      "get": {
        "tags": ["Review"],
        "operationId": "findAllReviewsByRecipe",
        "parameters": [
          {
            "name": "recipe-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{recipe-id}": {
      "get": {
        "tags": ["Recipe"],
        "operationId": "findRecipeById",
        "parameters": [
          {
            "name": "recipe-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/RecipeResponse" }
              }
            }
          }
        }
      }
    },
    "/recipes/author": {
      "get": {
        "tags": ["Recipe"],
        "operationId": "findAllRecipesByAuthor",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["KOREAN", "QUICK", "VEGAN", "BAKERY", "PASTA"]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseRecipeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/activate-account": {
      "get": {
        "tags": ["Authentication"],
        "operationId": "confirm",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    }
  },
  "components": {
    "schemas": {
      "ReviewRequest": {
        "required": ["comment", "recipeId"],
        "type": "object",
        "properties": {
          "rate": {
            "maximum": 5,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "comment": { "type": "string" },
          "recipeId": { "type": "integer", "format": "int64" }
        }
      },
      "CookingStep": {
        "required": ["number", "text"],
        "type": "object",
        "properties": {
          "text": { "type": "string" },
          "image": { "type": "string" },
          "number": { "type": "integer", "format": "int32" }
        }
      },
      "Ingredient": {
        "required": ["name", "quantity"],
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "quantity": { "type": "string" }
        }
      },
      "RecipeRequest": {
        "required": [
          "cookingSteps",
          "cookingTime",
          "description",
          "ingredientList",
          "preparationTime",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "preparationTime": {
            "type": "object",
            "properties": {
              "seconds": { "type": "integer", "format": "int64" },
              "zero": { "type": "boolean" },
              "nano": { "type": "integer", "format": "int32" },
              "negative": { "type": "boolean" },
              "positive": { "type": "boolean" },
              "units": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "durationEstimated": { "type": "boolean" },
                    "duration": {
                      "type": "object",
                      "properties": {
                        "seconds": { "type": "integer", "format": "int64" },
                        "zero": { "type": "boolean" },
                        "nano": { "type": "integer", "format": "int32" },
                        "negative": { "type": "boolean" },
                        "positive": { "type": "boolean" }
                      }
                    },
                    "timeBased": { "type": "boolean" },
                    "dateBased": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "cookingTime": {
            "type": "object",
            "properties": {
              "seconds": { "type": "integer", "format": "int64" },
              "zero": { "type": "boolean" },
              "nano": { "type": "integer", "format": "int32" },
              "negative": { "type": "boolean" },
              "positive": { "type": "boolean" },
              "units": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "durationEstimated": { "type": "boolean" },
                    "duration": {
                      "type": "object",
                      "properties": {
                        "seconds": { "type": "integer", "format": "int64" },
                        "zero": { "type": "boolean" },
                        "nano": { "type": "integer", "format": "int32" },
                        "negative": { "type": "boolean" },
                        "positive": { "type": "boolean" }
                      }
                    },
                    "timeBased": { "type": "boolean" },
                    "dateBased": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "cookingSteps": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CookingStep" }
          },
          "ingredientList": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Ingredient" }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["KOREAN", "QUICK", "VEGAN", "BAKERY", "PASTA"]
            }
          }
        }
      },
      "RegistrationRequest": {
        "required": ["email", "firstname", "lastname", "password"],
        "type": "object",
        "properties": {
          "firstname": { "type": "string" },
          "lastname": { "type": "string" },
          "email": { "type": "string" },
          "password": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "AuthenticationRequest": {
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": { "jwtToken": { "type": "string" } }
      },
      "PageResponseReviewResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReviewResponse" }
          },
          "number": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int64" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" }
        }
      },
      "ReviewResponse": {
        "type": "object",
        "properties": {
          "rate": { "type": "number", "format": "double" },
          "comment": { "type": "string" },
          "ownFeedback": { "type": "boolean" }
        }
      },
      "PageResponseRecipeResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RecipeResponse" }
          },
          "number": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int64" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" }
        }
      },
      "RecipeResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "preparationTime": {
            "type": "object",
            "properties": {
              "seconds": { "type": "integer", "format": "int64" },
              "zero": { "type": "boolean" },
              "nano": { "type": "integer", "format": "int32" },
              "negative": { "type": "boolean" },
              "positive": { "type": "boolean" },
              "units": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "durationEstimated": { "type": "boolean" },
                    "timeBased": { "type": "boolean" },
                    "dateBased": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "cookingTime": {
            "type": "object",
            "properties": {
              "seconds": { "type": "integer", "format": "int64" },
              "zero": { "type": "boolean" },
              "nano": { "type": "integer", "format": "int32" },
              "negative": { "type": "boolean" },
              "positive": { "type": "boolean" },
              "units": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "durationEstimated": { "type": "boolean" },
                    "timeBased": { "type": "boolean" },
                    "dateBased": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "cookingSteps": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CookingStep" }
          },
          "ingredientList": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Ingredient" }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["KOREAN", "QUICK", "VEGAN", "BAKERY", "PASTA"]
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth descritption",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}